* Fixes Issue #9261 * Fixes Issue #9258 * Added static CurrentCulture property that relays to the instance

+++ Tests/AssemblyTest/TestDataClass.cs
@@ -46,8 +46,8 @@
             set { countryDE = value; OnNotifyPropertyChanged("CountryDE"); }
         }
         
-        private int area = 0;
-        public int Area
+        private double area = 0;
+        public double Area
         {
             get { return area; }
             set { area = value; OnNotifyPropertyChanged("Area"); }
